fix: Don't send CSRF unneccessarily on unmodifying requests
