+++ shared/helpers/AuthenticationHelper.test.ts
